pyproject.toml
[tool.poetry]
name = "fast-zero"
version = "0.1.0"
description = ""
authors = ["Paulo Jalowyj <paulo.jalowyj@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = { extras = ["standard"], version = "^0.112.2" }
pydantic = {extras = ["email"], version = "^2.8.2"}
sqlalchemy = "^2.0.34"
pydantic-settings = "^2.4.0"
alembic = "^1.13.2"
pyjwt = "^2.9.0"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.3"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"

[tool.taskipy.tasks]
run = 'fastapi dev fast_zero/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fast_zero -vv'
post_test = 'coverage html'
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
